requirements.txt
slack-sdk==3.23.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
google-generativeai==0.3.2
python-dotenv==1.0.0
flask==2.3.3
gunicorn==21.2.0
requests==2.31.0
pyngrok==7.0.0